+++ RememberRoot.hpp
@@ -18,7 +18,7 @@ bool Remember::Root::isUnsynced()
    return (synced == false);
 }
 
-size_t Remember::Root::compileDataSize() const
+uint64_t Remember::Root::compileDataSize() const
 {
    return get().size();
 }
@@ -36,7 +36,7 @@ void Remember::Root::set(const DataVector& data)
    if (0 == data.size())
       return;
 
-   size_t cursor = 0;
+   uint64_t cursor = 0;
    read(data, cursor);
 }
